feat(ios): Starts TestUtils 'namespace' for testing funcs
