+++ ios/engine/KMEI/KeymanEngineTests/TestUtils/TestUtils.swift
@@ -1,0 +1,16 @@
+//
+//  TestUtils.swift
+//  KeymanEngineTests
+//
+//  Created by Joshua Horton on 2/19/20.
+//  Copyright © 2020 SIL International. All rights reserved.
+//
+
+import Foundation
+
+/**
+ * This class will 'namespace' the various testing utility functions used for our unit tests.
+ */
+class TestUtils {
+
+}
